🏗️ Architect: Modularized skymap generation logic
Refactored the skymap plotting logic within `apts/plotting/` into a dedicated `skymaps` sub-package to improve maintainability and follow the Single Responsibility Principle.

Key changes:
- Created `apts/plotting/skymaps/` package with a comprehensive `__init__.py` for the internal API.
- Extracted target resolution logic (searching across Messier, NGC, Stars, and Planets) into `resolver.py`.
- Relocated specialized generators (`skymap_polar.py`, `skymap_zoom.py`, `skymap_texture.py`) into the new package.
- Modularized shared coordinate system and axes setup logic into `utils.py`.
- Refactored `apts/plotting/skymap.py` into a lean dispatcher.
- Updated all internal imports and unit test mocks to reflect the new structure.

All plotting and observation tests passed successfully.
